DX12: improved linearizeDepth01, slice from depth

+++ DX12/D3D12Research/D3D12/Core/StringHash.h
@@ -74,9 +74,9 @@ namespace std
 	template <>
 	struct hash<StringHash>
 	{
-		uint32_t operator()(const StringHash& hash) const
+		size_t operator()(const StringHash& hash) const
 		{
-			return hash.m_Hash;
+			return (size_t)hash.m_Hash;
 		}
 	};
 }
